refactor(): type 패키지로 분리
