+++ app/src/main/java/nextstep/jwp/http/request/RequestLine.java
@@ -1,6 +1,7 @@
 package nextstep.jwp.http.request;
 
 import com.google.common.base.Strings;
+import nextstep.jwp.http.request.type.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
